Simplify CONTRIBUTING.md to be just guidelines
We don't need to include too much here because the real information for
contributing is in the handbook. This page is a standard page for Github
repos, so trimming it down to just a few links to sections in the
handbook.

Plus the policies for code of conduct and GPL.